refactor: isolate the nested-sampling algorithm
