+++ cicd/tests/test_phantom_conditioning.py
@@ -9,7 +9,7 @@ from jax import numpy as jnp
 import jaxns.phantom_eval as jax_phantom
 import jaxns.phantom_eval_ref as ref_phantom
 from jaxns import shrinkage
-from jaxns.race_tree import BlockState
+from jaxns.algorithm.race_tree import BlockState
 
 
 class PerClusterCountFixture(NamedTuple):
