+++ src/jaxns/algorithm/race_tree.py
