+++ src/jaxns/shrinkage.py
@@ -8,7 +8,7 @@ from jax.scipy.special import logsumexp
 from jaxns.log_semiring import LogSpace, normalise_log_space
 from jaxns.mixed_precision import mp_policy
 from jaxns.pytree import PureDataclassPytree
-from jaxns.race_tree import BlockState
+from jaxns.algorithm.race_tree import BlockState
 from jaxns.types import BoolArray, FloatArray, PRNGKey
